+++ src/store/reducer.ts
@@ -1,0 +1,16 @@
+import * as actionTypes from "./actionTypes";
+
+const initialState: PostState = {
+  postList: [],
+  isLoading: false,
+  error: null,
+};
+
+const reducer = (state = initialState, action: PostAction): PostState => {
+  switch (action.type) {
+    default:
+      return state;
+  }
+};
+
+export default reducer;